backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
pinecone-client
